remedied loop by extracting topArtists method into userServiceImpl
also addressed hibernates issues with resetting/instantiating a collection in an objects property field
