+++ Backend/src/main/resources/application-local.properties
@@ -1,0 +1,18 @@
+spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
+spring.datasource.url = jdbc:mysql://localhost:3306/soundoff
+spring.datasource.username = haydenhuffman
+spring.datasource.password = 0xoxmxox0
+
+
+
+# none, create, create-drop, update
+spring.jpa.hibernate.ddl-auto = update
+spring.jpa.show-sql = true
+
+spring.thymeleaf.cache = false
+
+spring.servlet.multipart.max-request-size=10MB
+spring.servlet.multipart.max-file-size=10MB
+
+token.expiration=600000
+token.refreshExpiration=2400000
