+++ Backend/src/main/resources/application.properties
@@ -1,15 +1,1 @@
-spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
-spring.datasource.url = jdbc:mysql://roundhouse.proxy.rlwy.net:24269/railway
-spring.datasource.username = root
-
-# none, create, create-drop, update
-spring.jpa.hibernate.ddl-auto = update
-spring.jpa.show-sql = true
-
-spring.thymeleaf.cache = false
-
-spring.servlet.multipart.max-request-size=10MB
-spring.servlet.multipart.max-file-size=10MB
-
-token.expiration=600000
-token.refreshExpiration=2400000
+spring.profiles.active=local
